quimby: rely on Github Actions instead of GGG
With dd/ci-swap-azure-pipelines-with-github-actions, there's no need to
create PRs we don't intend to close later in GGG. We can run a more
robust set of tests on our own forks instead.

This does require us to set a Quimby-specific config. For now, assume
the user made it themselves.

The code becomes significantly cleaner with the loss of 'gh', but
without a hook around git format-patch, Quimby can't yet die. Oh, well.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>