quimby: save state of version; autogen rangediff
Rangediffs help reviewers check whether their comments were addressed or
not, but in order to generate them, the code author needs to remember to
save a reference to the earlier version of their work.

Some people, like me, have crappy memory - so autosave each version in
an unlikely-to-collide branch.

This way we can also autogenerate a rangediff if we happen to have that
branch format from the previous version.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>